Creating Index.html
Creating initial setup for index.html. Setting up the boostrap stylesheet. Along with this is I have linked the css stylesheet that I have created for this page and other pages  and named it home.css. Then I have created the initial set up for the navigation bar. Used unordered list and then listing the links that I am used for the navigation bar.