Style for picture of Home page
Setting up the style for the picture on the homepage. Setting the border to 1px solid black, set the radius to 4px, the padding to 5px, and finally the width of the picture to 150px. 
Next when you hover over the image to add a shadow effect when you hover over the picture.