Part 3
In this section I used the p class tag to add a backcolor to the text "this is the homepage". I then added more text to the page and set the font family to courier, the color of the font blue, and the font size to 30px. I then linked an image and made the image a thumbnail and also set the width to 200px. Finally I added attribution and the sources that I used to create the page.