Part 2 of code
Using the boostrap script src and linking it to the page. Setting up the body background color and then using page breaks to organize the page. Then finally setting up the specifications for the picture that will be shown on the page. Setting the border of the image to be 1px and border color to be black. Then setting the border radius to be 4px, padding to be 5px and the width to be 150px. After this I used an image hover to have a shadow on the image when it is hovered over. Finally closing the style and head