Final Steps of the page
Opening the body tag. Using the p class tag to set a backcolor for the text "this is the homepage" as we were supposed to do for this assignment. Then linking an image and creating a thumbnail for the image that opens a larger image when it is clicked on. Finally closing the body tag and linking the sources that I used to create the page and posting them on this page.